main.py
import init
import torch, diffusers, math
import numpy as np
from PIL import Image
from transformers import CLIPTextModel, CLIPTokenizer, CLIPImageProcessor
from diffusers import (UNet2DConditionModel, 
                    AutoencoderKL, 
                    StableDiffusionPipeline, 
                    StableDiffusionImg2ImgPipeline,
                    EulerDiscreteScheduler)
from transformers import CLIPSegProcessor, CLIPSegForImageSegmentation
import matplotlib.pyplot as plt
import torch.nn.functional as F



def main():
    print("Running main application...")
    # Your main application logic here

if __name__ == "__main__":
    main()