requirements.txt
diffusers
transformers
accelerate
xformers
pillow
matplotlib
scipy
torchmetrics
